Redesigned replies, added reply functionality, and added replies to getPost asynchronously.
Forgive my lovecraftian asynch functions, I will probably forget how it works later on so I'm gonna write everything I remember.

In the getPost funciton for the controller, we have two functions, loadCommentReps and getRepliesImgs.

After loading the necesary user, post, and log details for the full_post, we delve into a db.findMany to grab all the comments.

Through this, we call loadCommentReps to asynchronously prepare each comment. This function prepares a promise wherein it resolves after completing its own loops.

It aims to find a user to grab the profile picture AND THEN grabs an array of replies and calls getRepliesImgs.

In this next function, it also finds a user each loop to place their profile picture. If the loop is done, it returns its promise.